winrequirements.txt
click==8.1.7 
customtkinter==5.2.2
libusb1==3.1.0
pyusb==1.3.1 
pywin32==308
packaging==24.1
pillow==10.4.0 
Svg.path==6.3
tqdm==4.66.5
lxml==5.3.0
cairosvg==2.7.1